Refactor deployment scripts
Why:

As the number of experiments was increasing designing and deploying each
of them was getting cumbersome and disorganized.

How:

Centralize all save/deploy into a single script deploy.py. This scripts
has two commands, register and schedule. It supports different types of
jobs (yet experiments and analysis).

register.sh and schedule.sh helps keeping the interface simple despite
being forced to execute the registration phase and the scheduling phase
in different environments (first one is in container, second one is
directly on the cluster).