Support values with | chars ...
Why:

The escaping of command line arguments by singularity is breaking the
interface, making it impossible to support array arguments with quoted
items when containing points.

How:

Beware, that's ugly. Make the argument array-like, but join it back with
`" ".join(values)` to then split it on `|`s. This way we support arrays
with spaces inside items... At least it works... :(