Fix optimizer checkpointing
Why:

During checkpointing only the model was saved. That means each time we
resumed training the optimizer was reset to original state, with
momentum buffer's being wiped out. Also, analysis using the optimizer
were not valid since 1) the optimizer was not loaded from checkpoint 2)
it was not reset at every steps so momentum buffer was still
accumulating during analysis.

How:

1) Save both model and optimizer state_dict when checkpoiting
2) Combine both model and optimizer in analysis when restoration is required.