Add truncated PR
Why:

Most models are too large to be able to compute even the block-diagonal
PR. We used to subsample the parameters but that may miss a large
portion of the correlations. A possibly better solution would be to use
sketching, or truncated covariance/eigen values to compute the PR.
We know for sure that the estimation is lower
than the true PR based on tests, but it may be still higher than the
subsampled version as the random projection of the truncated version
should capture most of the strong correlations. For the subsampled
version, it is hit or miss.

How:

For each layers, compute the truncated covariance and use it to compute
an estimation of the PR.