Add better support for batch-norm in PR analyses
Why:

When computer other analysis such as Fisher-Rao norm, we compute
gradients of each sample taking into account the effect of batch-norm
over mini-batches. We should do the same for the PR-based analyses.

How:

Compute output over mini-batch with same size as used to train model and
then compute the loss only over the first K samples, the mini-batch size
of the analysis. This way we can compute fisher-like PR(sgd-param) with
batch-size 1.