Add bin/deploy to new flow-execute behavior
Why:

Config files were binded on container at /config. This was problematic
because the standard path for configuration files is different, thus we
needed to manually point to those config files in commandlines. In the
most recent version, flow-execute copies .config dir in the container's home.
We can now drop all references to /config dir.