Center PR metrics with params specifically
Why:

The total centering is not very well justified theoretically, while
centering by subtracting the original parameter values is. We want to
analyze the movement from this point, not around the center of the
movements or from the origin (0).

Also, we may want to understand the effect of the variation of norms
across different steps. To enable this, we add a normalization options
to the PR metric.