Add BD-Fisher-PR
Why:

The function approximation in BD-PR is very crude, using a few samples
for estimation. We could rather use the matrix F to estimate our
function via F theta. It is prohibitively expensive, but it would be a
very important comparison to the validity of the crude function estimation.

How:

As for BD-PR, compute a cloud of new functions, but then compute the
KFAC approximation of F and the projection block-wise. The participation
ratio is then computed on the projections.