Properly count number of classes
Why:

There is some datasets where number of classes is greater than size of
minibatches. In this case the estimated number of classes based on first
mini-batch is wrong and it leads to the error:
```
Run timeError: cuda runtime error (59) : device-side assert triggered [...]
```

How:

Loop over a few mini-batches (20 for now) to make sure we have most
likely to have seen all of the classes, assuming ratio
batch_size/number_of_classes is not to close to 0. and that classes
are balanced.