Set num_classes manually
Related: 659e4d7385dcdef6ea93d60505de3f11ad85b8da

Why:

Some datasets have many classes and looking at a few batches is not
enough to get the entire set of classes. There can be no
heuristic for the number of batches to look at for a complete set of
classes. Setting it manually in datasets where that occurs is easier
and more efficient.