Add dataset wrapper for HDF5
Why:

HDF5 is needed for large datasets like ImageNet where the data is
scattered across many files.

Note:

HDF5 dataset has a lazy instantiation of the HDF5 file. This is
necessary for use cases with many workers, as the lazy instantiation
results in opening the file in the workers rather than in the main
process and copying the file handler. The latter is not thread-safe.