Fix convert_params bug in asha
Why:

The conversion was done when 'data' was present in params, but it should be the
opposite. When 'data' is present, it means `params` is the argument of a
registered task, hence it does no required any conversion. It is when
'data' is absent that `params` is a sample from ASHA and needs
conversion.