Fix (again) convert_params bug in asha
Why:

ASHA does not return the entire dictionnary of arguments of the task
when promoting but rather the subset corresponding to the optimized
space. Rather than looking at irrelevant fields such as 'data', we
should look at the type of `optimizer.lr_scheduler.milestones` which is
the field we actually convert. If it is a list, then it is a registered
task's arguments, otherwise it is a sample from ASHA.