Cancel duplicates in ASHA
Why:

Two workers may try to sample at the same time. If they are
both promoting they will likely end-up promoting twice the
same configuration, ending up with 2 identical trials. This is very
problematic as there is just a few promotions and duplicates are harmful
for the diversity of solutions.

How:

After registration, look at the other trials to see if there is any
duplicates. If there is, only conserve the first trial registered and
cancel the others. No need to re-sample to compensate, there is a
`create_trial` task registered at the end anyway.