Add registration of final trials
Why:

At end of HPO, we will register 20 trial per config to get an evaluation
of the distribution of the performances for the best set of HPs. The
configurations are those in final rung with best performance, worst
performance and mean of all HPs. That makes it 60 trials per (dataset,
model) pairs.