Compute checkpointing path as beginning of train
Why:

There is a problem with connection timeout with the client connection
because there is no queries for 5 minutes after the creation of the
connection.

How:

Create the client at beginning and compute the file path immediately,
discarding at the same time the client connection as it is now necessary
anymore.