Change MultiStepLR for ReduceLROnPlateau
Why:

It reduces the number of hyper-parameters to optimize and gives good
results based on literature.

Note that ReduceLROnPlateau is also problematic for HPO methods like
ASHA as the validation accuracy could suddenly drop after the max number
of epochs due to a sudden learning rate decay. Anyhow, gradual LR schedules
seems to be less performant based on the literature so we will avoid
them.