Turn create_trial into a long sequential task
Why:

We would need rollback handling of task creation within a task to avoid
the potential problem of generating a new create_trial while the current
one gets interrupted and resumed. Simply interrupting the task at the
end to force delayed re-execution is simpler and also has the advantage
of avoiding cluttering the task queue history.