Force thaw on suspend to avoid starvation
Why:

With large number of trials there is often a starvation when the random
subset cannot be thawed.

How:

When suspending, run thaw for the entire set of trials to make sure the
system does not starve. In worst case, if all trials are suspended,
`create_trial` will detect it and attempt to resume all trials.