Turn AFP into DPD centralized server
Why:

The amount of data moving around would just not scale. Creating a server
makes it possible to only transfer data to a single place and even
better we can run it on the private network of the DB to reduce network
I/O costs.

How:

For simplicity, two servers are building up metrics independently,
duplicating the work. Those servers can be limited to tags however,
making it possible to distribute the workload into a hashring of tags.