Adapt Dispatcher to support HPO
The parallel service needed to get passed full history before it call
suggest new HPs.

To support Mahler backend in the future, the task must be passed a
callback instead of a queue. The local Trial object will take care of
wrapping an observer fct to listen to mahler in fill the queue locally.
But the task itelf may me remote and cannot access the queue, hence why
it needs a callback definition to register data remotely.