Adapt COCOBenchmark to new dispatcher
There is no more need to embed the HPO process in the benchmark itself.
The benchmark will rather only define the function to optimise.

Benchmark class as been turned to standard class again as it requires to
many methods to be supported by dataclasses. It is not a mere data
container.