Add timestamps to MahlerTrial
The timestamps are a bit tricky for mahler, because there is often a
significant delay between the suspend signal and the actual suspension
(~30s). To handle this, start and stop timestamps are rather computed
on-the-fly based on the history of status of the trials, combined with
the observations. For instance, the task may continue running for 30s
after the status was set to suspend, therefore `stop` timestamp must be
equal to the timestamp of the last observation, not the timestamp of the
Suspended signal. If we would use the latter as a reference, then there
would be observations after the `stop`, which is incoherent.